Heights using number of levels*constant-Halle
Compute heights using number of levels times some selected constant on the city Halle. 
The computed heights are then returned as rater GeoTiff data.
The total number of buildings =  40,636, but we have computed heights on only 8,440 buildings (around 20% because these are the only buildings that have the number of levels available).